import-graphql-files * Add .graphqlconfig to support graphql plugins in IntelliJ and VScode * Add result from inspection query in graphql-schemas * Add definition of ApiDirective. Supports all the endpoints defined in the multi endpoint ApolloClient (realtimeVehiclesService.ts) * Add graphql-tag/loader to webpack config to support importing .graphql files from js/ts * Extract VEHICLES_QUERY, VEHICLE_UPDATES_SUBSCRIPTION and VEHICLE_FRAGMENT to .graphql files. Import them in useRealTimeVehicleData.
