+++ .graphqlconfig
@@ -1,0 +1,83 @@
+{
+  "projects": {
+    "vehicles": {
+      "name": "vehicles",
+      "schemaPath": "graphql-schemas/vehicles.graphql",
+      "includes": [
+        "graphql-schemas/ApiDirective.graphql",
+        "graphql-schemas/vehicles.graphql",
+        "src/logic/use-realtime-vehicle-data/VehicleFragment.graphql",
+        "src/logic/use-realtime-vehicle-data/VehiclesQuery.graphql",
+        "src/logic/use-realtime-vehicle-data/VehicleUpdatesSubscription.graphql"
+      ],
+      "extensions": {
+        "endpoints": {
+          "vehicles": {
+            "url": "https://api.entur.io/realtime/v1/vehicles/graphql",
+            "headers": {
+              "user-agent": "JS GraphQL"
+            },
+            "introspect": true
+          }
+        }
+      }
+    },
+    "journey-planner-v2": {
+      "name": "journey-planner-v2",
+      "schemaPath": "graphql-schemas/journey-planner-v2.graphql",
+      "includes": [
+        "graphql-schemas/ApiDirective.graphql",
+        "graphql-schemas/journey-planner-v2.graphql"
+      ],
+      "extensions": {
+        "endpoints": {
+          "journey-planner-v2": {
+            "url": "https://api.entur.io/journey-planner/v2/graphql",
+            "headers": {
+              "user-agent": "JS GraphQL"
+            },
+            "introspect": true
+          }
+        }
+      }
+    },
+    "journey-planner-v3": {
+      "name": "journey-planner-v3",
+      "schemaPath": "graphql-schemas/journey-planner-v3.graphql",
+      "includes": [
+        "graphql-schemas/ApiDirective.graphql",
+        "graphql-schemas/journey-planner-v3.graphql"
+      ],
+      "extensions": {
+        "endpoints": {
+          "journey-planner-v3": {
+            "url": "https://api.entur.io/journey-planner/v3/graphql",
+            "headers": {
+              "user-agent": "JS GraphQL"
+            },
+            "introspect": true
+          }
+        }
+      }
+    },
+    "mobility": {
+      "name": "mobility",
+      "schemaPath": "graphql-schemas/mobility.graphql",
+      "includes": [
+        "graphql-schemas/ApiDirective.graphql",
+        "graphql-schemas/mobility.graphql"
+      ],
+      "extensions": {
+        "endpoints": {
+          "mobility": {
+            "url": "https://api.entur.io/mobility/v2/graphql",
+            "headers": {
+              "user-agent": "JS GraphQL"
+            },
+            "introspect": true
+          }
+        }
+      }
+    }
+  }
+}
