+++ .prettierignore
@@ -1,3 +1,4 @@
 dist
 lib
 vendor
+graphql-schemas
