+++ src/logic/use-realtime-vehicle-data/VehicleFragment.graphql
@@ -1,0 +1,30 @@
+fragment VehicleFragment on VehicleUpdate {
+    vehicleRef
+    codespace {
+        codespaceId
+    }
+    operator {
+        operatorRef
+    }
+    line {
+        lineName
+        lineRef
+    }
+    serviceJourney {
+        serviceJourneyId
+    }
+    direction
+    mode
+    lastUpdated
+    lastUpdatedEpochSecond
+    expiration
+    expirationEpochSecond
+    speed
+    heading
+    monitored
+    delay
+    location {
+        latitude
+        longitude
+    }
+}
