+++ src/logic/use-realtime-vehicle-data/useRealtimeVehicleData.ts
@@ -1,21 +1,19 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useApolloClient, useQuery } from '@apollo/client'
 import type { FetchResult } from '@apollo/client'
-import { Filter } from '../services/realtimeVehicles/types/filter'
-import { RealtimeVehicle } from '../services/realtimeVehicles/types/realtimeVehicle'
-import {
-    VEHICLES_QUERY,
-    VEHICLE_UPDATES_SUBSCRIPTION,
-} from '../services/realtimeVehicles/graphql'
+import { Filter } from '../../services/realtimeVehicles/types/filter'
+import { RealtimeVehicle } from '../../services/realtimeVehicles/types/realtimeVehicle'
 import {
     SWEEP_INTERVAL_MS,
     DEFAULT_FETCH_POLICY,
     BUFFER_SIZE,
     BUFFER_TIME,
-} from '../constants'
-import { useSettings } from '../settings/SettingsProvider'
-import { useStopPlacesWithLines } from './useStopPlacesWithLines'
-import { useVehicleReducer, ActionType } from './useRealtimeVehicleReducer'
+} from '../../constants'
+import { useSettings } from '../../settings/SettingsProvider'
+import { useStopPlacesWithLines } from '../useStopPlacesWithLines'
+import { useVehicleReducer, ActionType } from '../useRealtimeVehicleReducer'
+import VEHICLE_UPDATES_SUBSCRIPTION from './VehicleUpdatesSubscription.graphql'
+import VEHICLES_QUERY from './VehiclesQuery.graphql'
 
 interface Return {
     realtimeVehicles: RealtimeVehicle[] | undefined
