+++ types/graphql.d.ts
@@ -1,0 +1,6 @@
+declare module '*.graphql' {
+    import { DocumentNode } from 'graphql'
+    const Schema: DocumentNode
+
+    export = Schema
+}
